backend/requirements.txt
Django==1.8
django-cors-headers==1.0.0
djangorestframework==3.1.1
Pillow==2.8.1
six==1.9.0

